It lives
Ripped out the `worker.py` bits from the [fo_api](#link?) repo to be a
standalone thing. (As it should be). Still doesn't do much more than
sleep and then report that its done, but the messaging protocol is
becoming more concrete. Time to check in so @Jon can replicate an
environment.